Fixed compatibility with latest version of mCRL2.

+++ src/ParityGame_pbes.h
@@ -28,8 +28,8 @@ void ParityGame::assign_pbes( mcrl2::pbes_system::pbes<Container> &pbes,
     verti begin = 0, end = 3;
     for (verti v = begin; v < end; ++v)
     {
-        std::set<unsigned> deps = pgg.get_dependencies(v);
-        for ( std::set<unsigned>::const_iterator it = deps.begin();
+        std::set<long unsigned> deps = pgg.get_dependencies(v);
+        for ( std::set<long unsigned>::const_iterator it = deps.begin();
                 it != deps.end(); ++it )
         {
             verti w = (verti)*it;
